+++ tests/scheduler/backends/qblox/operation_handling/test_acquisitions.py
@@ -43,6 +43,7 @@ from quantify_scheduler.instrument_coordinator.components.qblox import (
     QbloxInstrumentCoordinatorComponentBase,
 )
 from quantify_scheduler.operations.acquisition_library import SSBIntegrationComplex
+from quantify_scheduler.operations.control_flow_library import Loop
 from quantify_scheduler.operations.gate_library import Measure
 from quantify_scheduler.operations.pulse_library import SquarePulse
 from quantify_scheduler.resources import ClockResource
@@ -2047,6 +2048,114 @@ def test_append_measurements(
                 [[2 + 3j, 4 + 5j], [6 + 7j, 8 + 9j], [10 + 11j, 12 + 13j]],
                 coords={"acq_index_1": [0, 1]},
                 dims=["repetition", "acq_index_1"],
+            ),
+        }
+    )
+
+    xr.testing.assert_equal(data, expected_dataset)
+
+    instr_coordinator.remove_component("ic_cluster0")
+
+
+def test_looped_measurements(
+    mock_setup_basic_transmon, make_cluster_component
+):  # pylint: disable=too-many-locals
+    hardware_cfg = {
+        "config_type": "quantify_scheduler.backends.qblox_backend.QbloxHardwareCompilationConfig",
+        "hardware_description": {
+            "cluster0": {
+                "instrument_type": "Cluster",
+                "ref": "internal",
+                "modules": {
+                    "3": {"instrument_type": "QRM"},
+                },
+            }
+        },
+        "hardware_options": {
+            "modulation_frequencies": {
+                "q0:res-q0.ro": {
+                    "interm_freq": 50e6,
+                }
+            }
+        },
+        "connectivity": {
+            "graph": [
+                ["cluster0.module3.complex_output_0", "q0:res"],
+            ]
+        },
+    }
+
+    # Setup objects needed for experiment
+    mock_setup = mock_setup_basic_transmon
+    ic_cluster0 = make_cluster_component("cluster0")
+
+    instr_coordinator = mock_setup["instrument_coordinator"]
+    instr_coordinator.add_component(ic_cluster0)
+
+    quantum_device = mock_setup["quantum_device"]
+    quantum_device.hardware_config(hardware_cfg)
+
+    q0 = mock_setup["q0"]
+    q0.clock_freqs.readout(50e6)
+
+    # Define experiment schedule
+    schedule = Schedule("test multiple measurements", repetitions=2)
+    schedule.add(
+        Measure(
+            "q0",
+            acq_index=0,
+            acq_protocol="SSBIntegrationComplex",
+            bin_mode=BinMode.APPEND,
+        ),
+        control_flow=Loop(3),
+    )
+
+    # Change acq delay, duration and channel
+    q0.measure.acq_delay(1e-6)
+    q0.measure.integration_time(5e-6)
+    q0.measure.acq_channel(0)
+
+    # Setup dummy acquisition data
+    ic_cluster0.instrument.set_dummy_binned_acquisition_data(
+        slot_idx=3,
+        sequencer=0,
+        acq_index_name="0",
+        data=[
+            DummyBinnedAcquisitionData(data=(10000, 15000), thres=0, avg_cnt=0),
+            DummyBinnedAcquisitionData(data=(20000, 25000), thres=0, avg_cnt=0),
+            DummyBinnedAcquisitionData(data=(30000, 35000), thres=0, avg_cnt=0),
+            DummyBinnedAcquisitionData(data=(40000, 45000), thres=0, avg_cnt=0),
+            DummyBinnedAcquisitionData(data=(50000, 55000), thres=0, avg_cnt=0),
+            DummyBinnedAcquisitionData(data=(60000, 65000), thres=0, avg_cnt=0),
+        ],
+    )
+
+    # Generate compiled schedule
+    compiler = SerialCompiler(name="compiler")
+    compiled_sched = compiler.compile(
+        schedule=schedule, config=quantum_device.generate_compilation_config()
+    )
+
+    # Upload schedule and run experiment
+    instr_coordinator.prepare(compiled_sched)
+    instr_coordinator.start()
+
+    with pytest.warns(
+        FutureWarning,
+        match="The format of acquisition data of looped measurements in APPEND mode will change in quantify-scheduler>=0.18.0",
+    ):
+        data = instr_coordinator.retrieve_acquisition()
+
+    instr_coordinator.stop()
+
+    # Assert intended behaviour
+    assert isinstance(data, Dataset)
+    expected_dataset = Dataset(
+        {
+            0: DataArray(
+                [[2 + 3j, 4 + 5j, 6 + 7j], [8 + 9j, 10 + 11j, 12 + 13j]],
+                coords=None,
+                dims=["repetition", "loop_repetition"],
             ),
         }
     )